refactor: DatabaseCleanUp을 DatabaseCleaner로 변경
